+++ packages/annotation/src/markers/index.tsx
@@ -1,4 +1,3 @@
 export * from './ArrowMarker'
 export * from './BluntMarker'
-export * from './LinearGradient'
 export * from './types'