+++ packages/annotation/src/markers/types.ts
@@ -1,4 +1,4 @@
-import { CssProps, NumericPositionSpec } from '../general'
+import { CssProps } from '@chsk/core'
 
 interface BaseMarkerProps {
     /** identifier */
@@ -19,17 +19,4 @@ export interface ArrowMarkerProps extends BaseMarkerProps {
 export interface BluntMarkerProps extends BaseMarkerProps {
     /** variant of arrowhead */
     variant: 'Circle' | 'Square' | 'Diamond'
-}
-
-export interface LinearGradientProps {
-    /** identifier for the gradient */
-    id: string
-    /** start position */
-    start: NumericPositionSpec
-    /** end position */
-    end: NumericPositionSpec
-    /** color stops */
-    stops: string[]
-    /** offsets for color stops */
-    offsets?: number[]
 }